CT evaluation works well, typer requires some work
Something is going wrong while casting tuples to structs that
manifests as a unification error.

Otherwise, CT evaluation (even returning structs) works extremely
well.

When the typer is relaxed (introducing other issues, but allowing
us to test CT eval), structs (which are translated into tuples)
can be passed as arguments and returned.

The next major problem to solve is mem-windows in compile-time eval,
introducing mutability. This will be followed by an IR that can
be accessed by subsequent passes.